Organize the classes into suitable java packages, duke.command, duke.task, duke.exception
