+++ src/main/data/duke.txt
@@ -1,2 +1,2 @@
-T | 1 | add book
 T | 1 | read book
+T | 1 | new book
