+++ src/main/java/duke/Parser.java
@@ -1,3 +1,11 @@
+package duke;
+
+import duke.command.*;
+import duke.exception.DukeException;
+import duke.task.Deadline;
+import duke.task.Event;
+import duke.task.Todo;
+
 public class Parser {
 
     public static Command parse(String fullCommand) throws DukeException {
