+++ src/main/java/duke/command/DeleteCommand.java
@@ -1,16 +1,24 @@
-public class AddCommand extends Command {
+package duke.command;
 
-    Task t;
+import duke.Storage;
+import duke.Ui;
+import duke.exception.DukeException;
+import duke.task.Task;
+import duke.task.TaskList;
 
-    AddCommand(Task t) {
-        this.t = t;
+public class DeleteCommand extends Command {
+
+    private int index;
+
+    public DeleteCommand(int index) {
+        this.index = index;
     }
 
     @Override
     public void execute(TaskList tasks, Ui ui, Storage storage) throws DukeException {
-        tasks.addTask(t);
 
-        System.out.println("Got it. I've added this task:");
+        Task t = tasks.removeTask(index);
+        System.out.println("Noted. I've removed this task:");
         System.out.println(t);
         if (tasks.getSize() == 1) {
             System.out.println("Now you have 1 task in the list.");
