+++ src/main/java/duke/command/DoneCommand.java
@@ -1,8 +1,16 @@
+package duke.command;
+
+import duke.Storage;
+import duke.Ui;
+import duke.exception.DukeException;
+import duke.task.Task;
+import duke.task.TaskList;
+
 public class DoneCommand extends Command {
 
     private int index;
 
-    DoneCommand(int index) {
+    public DoneCommand(int index) {
         this.index = index;
     }
