+++ src/main/java/duke/command/UnknownCommand.java
@@ -1,6 +1,13 @@
+package duke.command;
+
+import duke.Storage;
+import duke.Ui;
+import duke.exception.DukeException;
+import duke.task.TaskList;
+
 public class UnknownCommand extends Command {
 
-    UnknownCommand() {}
+    public UnknownCommand() {}
 
     @Override
     public void execute(TaskList tasks, Ui ui, Storage storage) throws DukeException {
