+++ src/main/java/duke/exception/DukeException.java
@@ -1,8 +1,10 @@
+package duke.exception;
+
 public class DukeException extends Exception{
 
     String errorDescription;
 
-    DukeException(String errorDescription) {
+    public DukeException(String errorDescription) {
         super(errorDescription);
         this.errorDescription = errorDescription;
     }
