+++ src/main/java/duke/task/Deadline.java
@@ -1,3 +1,5 @@
+package duke.task;
+
 public class Deadline extends Task {
     protected String by;
 
@@ -6,7 +8,7 @@ public class Deadline extends Task {
         this.by = by;
     }
 
-    Deadline(String description, String by, boolean isDone) {
+    public Deadline(String description, String by, boolean isDone) {
         super(description, isDone);
         this.by = by;
     }
