+++ src/main/java/duke/task/TaskList.java
@@ -1,3 +1,6 @@
+package duke.task;
+import duke.exception.DukeException;
+
 import java.util.ArrayList;
 import java.util.List;
 
@@ -5,11 +8,11 @@ public class TaskList {
 
     private List<Task> tasks;
 
-    TaskList() {
+    public TaskList() {
         this.tasks = new ArrayList<>(100);
     }
 
-    TaskList(List<Task> tasks) {
+    public TaskList(List<Task> tasks) {
         this.tasks = tasks;
     }
