+++ src/test/java/StorageTest.java
@@ -1,3 +1,4 @@
+import duke.Storage;
 import org.junit.jupiter.api.Test;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
